remove .env and add firebase data to env
