--- .env
@@ -1,2 +1,0 @@
-VITE_API_BASE=https://bridge-ai-backend-cagfd2exh5h6gwgs.westeurope-01.azurewebsites.net
-VITE_DEFAULT_REDIRECT=/dashboard
